Rename ``vcgutils.py`` to ``vcg_printer.py``
